server/requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic[email]
python-jose[cryptography]  # For JWT tokens
passlib[bcrypt]  # For password hashing